refactor: centralise types in folder
