+++ src/electron.d.ts
@@ -1,36 +1,12 @@
 import type { Voice } from './types/voice';
-
-export interface Slide {
-    index: number;
-    image: string;
-    src: string; // The full path/url
-    notes: string;
-}
-
-export interface ConvertResponse {
-    success: boolean;
-    slides: Slide[];
-    error?: string;
-}
-
-export interface BasicElectronResult {
-    success: boolean;
-    error?: string;
-}
-
-export interface SlidesElectronResult extends BasicElectronResult {
-    slides?: Slide[];
-}
-
-export interface VideoElectronResult extends BasicElectronResult {
-    outputPath?: string;
-}
-
-export interface SlideAudioEntry {
-    index: number;
-    sectionIndex?: number;
-    audioData: Uint8Array;
-}
+import type {
+    BasicElectronResult,
+    ConvertResponse,
+    Slide,
+    SlideAudioEntry,
+    SlidesElectronResult,
+    VideoElectronResult,
+} from './types/electron';
 
 declare global {
     interface Window {
