+++ src/types/notes.ts
@@ -1,0 +1,4 @@
+export interface NoteSection {
+    speaker: string;
+    text: string;
+}
